Moves the Integrated Container logic up to TestModule
