+++ lib/ember-test-helpers/test-module-for-integration.js
@@ -4,6 +4,9 @@ import { getResolver } from './test-resolver';
 import { getContext, setContext } from './test-context';
 
 export default TestModule.extend({
+
+  isIntegration: true,
+
   init: function(name, description, callbacks) {
     this._super.call(this, name, description, callbacks);
     this.setupSteps.push(this.setupIntegrationHelpers);
@@ -53,24 +56,6 @@ export default TestModule.extend({
       self.actionHooks[actionName] = handler;
     };
 
-  },
-
-  setupContainer: function() {
-    var resolver = getResolver();
-    var namespace = Ember.Object.create({
-      Resolver: { create: function() { return resolver; } }
-    });
-
-    if (Ember.Application.buildRegistry) {
-      var registry;
-      registry = Ember.Application.buildRegistry(namespace);
-      registry.register('component-lookup:main', Ember.ComponentLookup);
-      this.registry = registry;
-      this.container = registry.container();
-    } else {
-      this.container = Ember.Application.buildContainer(namespace);
-      this.container.register('component-lookup:main', Ember.ComponentLookup);
-    }
   },
 
   setupContext: function() {
