+++ lib/ember-test-helpers/test-module.js
@@ -1,6 +1,8 @@
+import Ember from 'ember';
 import isolatedContainer from './isolated-container';
 import { getContext, setContext } from './test-context';
 import { Klass } from 'klassy';
+import { getResolver } from './test-resolver';
 
 export default Klass.extend({
   init: function(subjectName, description, callbacks) {
@@ -15,6 +17,10 @@ export default Klass.extend({
     this.description = description || subjectName;
     this.name = description || subjectName;
     this.callbacks = callbacks || {};
+
+    if (this.callbacks.integration) {
+      this.isIntegration = callbacks.integration;      
+    }
 
     this.initSubject();
     this.initNeeds();
@@ -97,7 +103,11 @@ export default Klass.extend({
   },
 
   setupContainer: function() {
-    this.container = isolatedContainer(this.needs);
+    if (this.isIntegration) {
+      this._setupIntegratedContainer();
+    } else {
+      this._setupIsolatedContainer();
+    }
   },
 
   setupContext: function() {
@@ -184,6 +194,30 @@ export default Klass.extend({
 
       })(keys[i]);
     }
+  },
+
+
+  _setupIsolatedContainer: function() {
+    this.container = isolatedContainer(this.needs);
+  },
+
+  _setupIntegratedContainer: function() {
+    var resolver = getResolver();
+    var namespace = Ember.Object.create({
+      Resolver: { create: function() { return resolver; } }
+    });
+
+    if (Ember.Application.buildRegistry) {
+      var registry;
+      registry = Ember.Application.buildRegistry(namespace);
+      registry.register('component-lookup:main', Ember.ComponentLookup);
+      this.registry = registry;
+      this.container = registry.container();
+    } else {
+      this.container = Ember.Application.buildContainer(namespace);
+      this.container.register('component-lookup:main', Ember.ComponentLookup);
+    }
   }
+
 });
